third_party/WebKit/LayoutTests/inspector/timeline/timeline-script-tag-1-expected.txt

CONSOLE MESSAGE: line 8: iframe loaded
Tests the Timeline API instrumentation of an HTML script tag.


ParseHTML
ParseHTML
----> EvaluateScript
--------> TimeStamp : SCRIPT TAG
--------> UpdateCounters
----> UpdateCounters
EvaluateScript Properties:
{
    children : <object>
    data : {
        lineNumber : <number>
        url : .../inspector/timeline/resources/timeline-iframe-data.html
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "EvaluateScript"
}
ParseHTML
ParseHTML
----> UpdateCounters
----> EventDispatch
--------> FunctionCall
------------> FunctionCall
------------> UpdateCounters
--------> UpdateCounters
----> UpdateCounters
----> MarkLoad
----> UpdateCounters
----> UpdateCounters
----> MarkDOMContent
----> UpdateCounters
